# res/openapi.yaml
openapi: 3.0.0
info:
  title: grommunio Admin API
  version: 1.0.0
servers:
  - url: '/api/v1'
    description: Default deployment path
paths:
...
  /system/logs:
    get:
      summary: Get available log files
      tags:
        - System Admin/Logs
      security:
        - JWTCookie: []
      responses:
        '200':
          description: List of log files returned
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      type: string
        '400':
          $ref: '#/components/responses/InvalidRequest'
        '500':
          $ref: '#/components/responses/ServerError'
  /system/logs/{file}:
    get:
      summary: Get log file content
      description: Returns the n last lines of the log file, similar to the `tail` command line tool
      tags:
        - System Admin/Logs
      security:
        - JWTCookie: []
      parameters:
        - name: file
          in: path
          required: true
          description: Name of the log file
          schema:
            type: string
        - name: n
          in: query
          description: Number of lines to return
          schema:
            type: integer
            default: 10
            minimum: 1
        - name: skip
          in: query
          description: Number of lines to skip
          schema:
            type: integer
            default: 0
            minimum: 0
        - name: after
          in: query
          description: Return all lines after given time. Overrides `n` and `skip`.
          schema:
            $ref: '#/components/schemas/precTime'
      responses:
        '200':
          description: List of log files returned
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      type: object
                      description: Log file entry
                      properties:
                        level:
                          type: integer
                          description: Priority level number as described in journalctl (1)
                        message:
                          type: string
                          description: Log message
                        time:
                          $ref: '#/components/schemas/precTime'
                        runtime:
                          type: number
                          description: Time since last reboot
        '400':
          $ref: '#/components/responses/InvalidRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServerError'
...
  /system/mconf/ldap:
    get:
      summary: Get the current LDAP configuration
      tags:
        - LDAP
        - System Admin/MConf
      security:
        - JWTCookie: []
      responses:
        '200':
          description: LDAP configuration returned
          content:
            application/json:
              schema:
                type: object
                properties:
                  ldapAvailable:
                    type: boolean
                    description: Whether ldap is currently active
                  data:
                    $ref: '#/components/schemas/ldapConfig'
        '400':
          $ref: '#/components/responses/InvalidRequest'
        '500':
          $ref: '#/components/responses/ServerError'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
    put:
      summary: Set new LDAP configuration
      tags:
        - LDAP
        - System Admin/MConf
      security:
        - JWTCookie: []
      parameters:
        - name: force
          description: Update configutation even if validation failed
          in: query
          schema:
            type: boolean
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ldapConfig'
      responses:
        '200':
          description: LDAP configuration updated
        '400':
          $ref: '#/components/responses/InvalidRequest'
        '500':
          $ref: '#/components/responses/ServerError'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
    delete:
      summary: Remove LDAP configuration and disable service
      tags:
        - LDAP
        - System Admin/MConf
      security:
        - JWTCookie: []
      responses:
        '200':
          description: LDAP deactivated
        '400':
          $ref: '#/components/responses/InvalidRequest'
        '500':
          $ref: '#/components/responses/ServerError'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
  /system/mconf/authmgr:
    get:
      summary: Get the current authmgr configuration
      tags:
        - System Admin/MConf
      security:
        - JWTCookie: []
      responses:
        '200':
          description: authmgr configuration returned
          content:
            application/json:
              schema:
                type: object
                properties:
                  authBackendSelection:
                    type: string
        '400':
          $ref: '#/components/responses/InvalidRequest'
        '500':
          $ref: '#/components/responses/ServerError'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
    put:
      summary: Set new authmgr configuration
      tags:
        - System Admin/MConf
      security:
        - JWTCookie: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                authBackendSelection:
                  type: string
                  enum: [always_mysql, always_ldap, externid]
      responses:
        '200':
          description: authmgr configuration updated
        '400':
          $ref: '#/components/responses/InvalidRequest'
        '500':
          $ref: '#/components/responses/ServerError'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
    delete:
      summary: Reset authmgr configuration to default
      tags:
        - System Admin/MConf
      security:
        - JWTCookie: []
      responses:
        '200':
          description: authmgr configuration reset
        '400':
          $ref: '#/components/responses/InvalidRequest'
        '500':
          $ref: '#/components/responses/ServerError'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
...
components:
  securitySchemes:
    JWTCookie:
      type: apiKey
      in: cookie
      name: grommunioAuthJwt
...
  schemas:
...
    precTime:
      type: string
      description: Precise date/time string including fractional seconds
      pattern: '^\d{4}-\d{2}-\d{2} \d{2}:\d{2}:\d{2}\.\d{1,6}$'
...
    ldapConfig:
      type: object
      properties:
        disabled:
          type: boolean
          description: Disable LDAP service
          default: False
        connection:
          type: object
          properties:
            server:
              type: string
              description: Space separated list of addresses of the LDAP server to connect to
            bindUser:
              type: string
              description: DN of the user to perform initial bind with
            bindPass:
              type: string
              description: Password for bindUser
            starttls:
              type: boolean
              description: Whether to initiate a StartTLS connection
        baseDn:
          type: string
          description: Base DN to use for user search
        objectID:
          type: string
          description: Name of an attribute that uniquely idetifies an LDAP object
        users:
          type: object
          description: Configuration for user search
          properties:
            username:
              type: string
              description: Name of the attribute that corresponds to the username (e-mail address)
            displayName:
              type: string
              description: Name of the attribute that contains the name
            filters:
              type: array
              description: List of additional filter expressions to use for user search
              items:
                type: string
                example: 'objectclass=person'
            filter:
              type: string
              description: LDAP search filter to apply to user lookup
              example: '(&(objectclass=person)(mailPrimaryAddress=*))'
            templates:
              type: array
              description: List of mapping templates to use
              items:
                type: string
                description: Name of a template defined in res/ldapTemplates.yaml
            attributes:
              type: object
              description: LDAP attribute -> PropTag mapping to used for LDAP import. Any mappings specified take precendence over active templates
              additionalProperties:
                type: string
                description: Name of the PropTag the atribute maps to
            defaultQuota:
              type: integer
              description: Storage quota of imported users if no mapping exists
            searchAttributes:
              type: array
              description: List of attributes to use for searching
              items:
                type: string
                description: LDAP attribute name
            aliases:
              type: string
              description: LDAP attribute containing alternative e-mail addresses
...
  responses:
    ServerError:
      description: An error occured while processing the request
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
                description: String representation of the exception
    InvalidRequest:
      description: Validation of input parameters failed
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
              errors:
                type: array
                description: List of errors encountered during validation
                items:
                  type: string
...
    ServiceUnavailable:
      description: One of the required external services is unavailable
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
                description: Description of the error
    NotFound:
      description: The requested resource could not be found
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
...